Moved product translation to product package
